+++ client/src/components/board/BoardIdeaCardContainer.js
@@ -4,9 +4,9 @@ import BoardInfoCard from "components/board/BoardInfoCard";
 import IdeaCard from "components/board/IdeaCard";
 import {SvgNotice} from "components/commons/SvgNotice";
 import AppContext from "context/AppContext";
-import React, {useContext, useEffect, useRef, useState} from 'react';
+import React, {useContext, useEffect, useState} from 'react';
 import {FaRegFrown} from "react-icons/fa";
-import InfiniteScroll from 'react-infinite-scroller';
+import InfiniteScroll from "react-infinite-scroll-component";
 import {UiLoadingSpinner} from "ui";
 import {UiCol} from "ui/grid";
 import {prepareFilterAndSortRequests} from "utils/basic-utils";
@@ -15,16 +15,11 @@ const BoardIdeaCardContainer = ({id, searchQuery}) => {
     const {user} = useContext(AppContext);
     const [ideas, setIdeas] = useState({data: [], loaded: false, error: false, moreToLoad: true});
     const [scrollTo, setScrollTo] = useState(null);
-    const isInitialMount = useRef(true);
+    const [page, setPage] = useState(0);
     useEffect(() => {
-        if (isInitialMount.current) {
-            isInitialMount.current = false;
-        } else {
-            setIdeas({...ideas, loaded: false});
-            onLoadRequest(1, true);
-        }
+        onLoadRequest(true);
         // eslint-disable-next-line
-    }, [id, searchQuery, user.localPreferences]);
+    }, [id, searchQuery, user.localPreferences.ideas]);
     const loadIdeas = () => {
         if (ideas.error) {
             return <SvgNotice Component={UndrawNoIdeas} title={<React.Fragment><FaRegFrown className={"mr-1"}/> Failed to load ideas</React.Fragment>}/>
@@ -36,10 +31,11 @@ const BoardIdeaCardContainer = ({id, searchQuery}) => {
             return <SvgNotice Component={UndrawNoIdeas} title={"No ideas yet."} description={"How about creating one?"}/>
         }
         return <InfiniteScroll
-            pageStart={0}
-            loadMore={page => onLoadRequest(page)}
+            style={{overflow: "hidden"}}
+            next={onLoadRequest}
             hasMore={ideas.moreToLoad}
-            loader={<UiCol className={"text-center mt-5 pt-5"} key={ideas.data.length}><UiLoadingSpinner/></UiCol>}>
+            dataLength={ideas.data.length}
+            loader={<UiCol className={"text-center mt-5 pt-5"}><UiLoadingSpinner/></UiCol>}>
             {ideas.data.map(ideaData => {
                 return <IdeaCard key={ideaData.id} ideaData={ideaData} onIdeaDelete={onIdeaDelete}/>
             })}
@@ -52,9 +48,10 @@ const BoardIdeaCardContainer = ({id, searchQuery}) => {
     const onIdeaDelete = (id) => {
         setIdeas({...ideas, data: ideas.data.filter(item => item.id !== id)});
     };
-    const onLoadRequest = (page, override = false) => {
+    const onLoadRequest = (override = false) => {
         const withQuery = searchQuery === "" ? "" : "&query=" + searchQuery;
-        return axios.get("/boards/" + id + "/ideas?page=" + (page - 1) + prepareFilterAndSortRequests(user.localPreferences.ideas) + withQuery).then(res => {
+        const currentPage = override ? 0 : page;
+        return axios.get("/boards/" + id + "/ideas?page=" + currentPage + prepareFilterAndSortRequests(user.localPreferences.ideas) + withQuery).then(res => {
             const data = res.data.data;
             data.forEach(element => element.tags.sort((a, b) => a.name.localeCompare(b.name)));
             if (override) {
@@ -62,6 +59,7 @@ const BoardIdeaCardContainer = ({id, searchQuery}) => {
             } else {
                 setIdeas({...ideas, data: ideas.data.concat(data), loaded: true, moreToLoad: res.data.pageMetadata.currentPage < res.data.pageMetadata.pages});
             }
+            setPage(currentPage + 1);
         }).catch(() => setIdeas({...ideas, error: true}));
     };
     useEffect(() => {