+++ CHANGELOG.md
@@ -33,6 +33,8 @@
 * improve api fetching response time with entitygraphs
 * force break-word on idea cards to avoid [this](https://cdn.discordapp.com/attachments/618782782348591111/796316504835751956/Screenshot_2021-01-06-10-57-17-570_org.mozilla.firefox.jpg)
 * on new webhook creation it should notify only created webhook not all board webhooks now
+* replaced react-infinite-scroller with react-infinite-scroll-component library and fixed common issues with ideas/comments loading
+such as race conditions and data not loading on components state update
 
 TODO dynamically load SVGs from assets/svg/undraw for smaller js chunk sizes